+++ R/CytoMethQC.R
@@ -30,6 +30,3 @@ gg <- plotMethLevels(methData, conf, share = T, save = save)
 gg <- plotCpGAnnotation(methData, hypo_hyper_def = c(0.2,0.8), conf, share = T, save = save)
 gg <- plotCntCommonCpG(methData, conf, save = save)
 
-
-
-